+++ src/controllers/extract.controller.js
@@ -1,10 +1,10 @@
 const fs = require("fs");
 const path = require("path");
 
-const convertPath = require("../../services/convertPath");
-const extractFramesFromVideo = require("../../services/extractFrames");
+const convertPath = require("../services/convertPath");
+const extractFramesFromVideo = require("../services/extractFrames");
 
-const { TEMP_DIR_VIDEOS, TEMP_DIR_FRAMES } = require("../../constants/paths");
+const { TEMP_DIR_VIDEOS, TEMP_DIR_FRAMES } = require("../constants/paths");
 
 exports.videoToFrame = async function (req, res, next) {
   try {
