+++ src/routes/videos.js
@@ -2,7 +2,7 @@ const express = require("express");
 
 const router = express.Router();
 
-const transferController = require("./controllers/transfer.controller");
+const transferController = require("../controllers/transfer.controller");
 
 router.post("/contents", transferController.transferMedia);
