+++ src/services/analyzeMovementRatio.js
