+++ src/services/convertFrameToBinaryArray.js
