+++ src/services/detectMovement.js
