+++ src/services/detectSingleShots.js
