+++ src/util/select1fpsFrames.js
