+++ src/util/validateArrays.js
